chore: move thread create and modify request types
