+++ src/leapfrogai_api/routers/openai/threads.py
@@ -2,10 +2,7 @@
 
 from fastapi import HTTPException, APIRouter, status
 from openai.types.beta import Thread, ThreadDeleted
-from leapfrogai_api.backend.types import ModifyThreadRequest
-from leapfrogai_api.routers.openai.requests.create_thread_request import (
-    CreateThreadRequest,
-)
+from leapfrogai_api.typedef.requests import ModifyThreadRequest, CreateThreadRequest
 from leapfrogai_api.data.crud_thread import CRUDThread
 from leapfrogai_api.routers.supabase_session import Session
 from leapfrogai_api.utils.validate_tools import (
