+++ src/leapfrogai_api/typedef/requests/__init__.py
@@ -2,11 +2,16 @@ from .auth_types import (
     CreateAPIKeyRequest as CreateAPIKeyRequest,
     ModifyAPIKeyRequest as ModifyAPIKeyRequest,
 )
-from .run_create_params_base import (
+
+from .run_create_base import (
     RunCreateParamsRequestBase as RunCreateParamsRequestBase,
 )
-from .run_create_params import RunCreateParamsRequest as RunCreateParamsRequest
+from .run_create import RunCreateParamsRequest as RunCreateParamsRequest
 from .run_modify import ModifyRunRequest as ModifyRunRequest
+
+from .thread_create import CreateThreadRequest as CreateThreadRequest
+from .thread_modify import ModifyThreadRequest as ModifyThreadRequest
+
 from .audio_types import (
     CreateTranscriptionRequest as CreateTranscriptionRequest,
     CreateTranscriptionResponse as CreateTranscriptionResponse,
